Move iterator benchmark into separate test
